requirements.txt
streamlit
python-dotenv
langchain-groq
pandas
pypdf
docx2txt
